% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fibonacci}
\alias{fibonacci}
\title{Simple rcpptimer example}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{vector giving integers for which to compute the Fibonacci sum}
}
\value{
vector of integers giving the Fibonacci sum for each element in
\code{n}
}
\description{
Time the computation of Fibonacci numbers
}
\details{
The function being timed is the following:

\code{int fib(int n) { return ((n <= 1) ? n : fib(n - 1) + fib(n - 2)); }}

Runtime for computations less than \code{n = 15} is nearly unmeasurable.
}
\examples{
\donttest{
fibonacci(n = rep(10*(1:4), 10))
# this function creates a global environment variable "times"
times
}
}
