% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rcpptimer.R
\name{print.rcpptimer}
\alias{print.rcpptimer}
\title{Print method for rcpptimer output}
\usage{
\method{print}{rcpptimer}(x, scale = TRUE, ...)
}
\arguments{
\item{x}{Object of class rcpptimer}

\item{scale}{Scale the timings and statistics to a more human readable format}

\item{...}{further arguments are ignored}
}
\description{
Prints the times object and scales the timings if appropriate.
If all timings are smaller than 1 microsecond, the timings are
printed in nanoseconds. If the smallest timing is higher than
a Millisecond / Seconds / Minutes / Hours, the timings are printed in
the unit of that threshold. This behavior can be disabled by setting
scale = FALSE.
}
