% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rcpptimer-package}
\alias{rcpptimer}
\alias{rcpptimer-package}
\title{rcpptimer: 'Rcpp' Tic-Toc Timer with 'OpenMP' Support}
\description{
To learn more about rcpptimer, start with the vignettes:
\code{browseVignettes(package = "rcpptimer")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rcpptimer.berrisch.biz}
  \item \url{https://github.com/BerriJ/rcpptimer}
  \item Report bugs at \url{https://github.com/BerriJ/rcpptimer/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Berrisch \email{Jonathan@Berrisch.biz} (\href{https://orcid.org/0000-0002-4944-9074}{ORCID})

}
\keyword{internal}
